feat: add feature flag for xwayland
