+++ src/client.c
@@ -243,52 +243,3 @@ void reset_floating_client_borders(int border_px)
         c->bw = border_px;
     }
 }
-
-bool wants_floating(struct client *c)
-{
-    if (c->type != X11_MANAGED && c->type != X11_UNMANAGED)
-        return false;
-
-    struct wlr_xwayland_surface *surface = c->surface.xwayland;
-    struct xwayland xwayland = server.xwayland;
-
-    if (surface->modal)
-        return true;
-
-    for (size_t i = 0; i < surface->window_type_len; ++i) {
-        xcb_atom_t type = surface->window_type[i];
-        if (type == xwayland.atoms[NET_WM_WINDOW_TYPE_DIALOG] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_UTILITY] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_TOOLBAR] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_POPUP_MENU] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_SPLASH]) {
-            return true;
-        }
-    }
-
-    struct wlr_xwayland_surface_size_hints *size_hints = surface->size_hints;
-    if (size_hints != NULL &&
-            size_hints->min_width > 0 && size_hints->min_height > 0 &&
-            (size_hints->max_width == size_hints->min_width ||
-            size_hints->max_height == size_hints->min_height)) {
-        return true;
-    }
-
-    return false;
-}
-
-bool is_popup_menu(struct client *c)
-{
-    struct wlr_xwayland_surface *surface = c->surface.xwayland;
-    struct xwayland xwayland = server.xwayland;
-    for (size_t i = 0; i < surface->window_type_len; ++i) {
-        xcb_atom_t type = surface->window_type[i];
-        if (type == xwayland.atoms[NET_WM_WINDOW_TYPE_POPUP_MENU] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_POPUP] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_MENU] ||
-                type == xwayland.atoms[NET_WM_WINDOW_TYPE_NORMAL]) {
-            return true;
-        }
-    }
-    return false;
-}
